[sdc] move `query`-module to `product`-subpackage
