+++ sdc/products/_query.py
